Moved charge reader.
