+++ source/databricks/calculation-engine/tests/calculation/wholesale/test_fee_calculators.py
@@ -24,7 +24,7 @@ from package.calculation.wholesale.fee_calculators import (
     get_count_of_charges_and_total_daily_charge_price,
 )
 from package.calculation.wholesale.wholesale_initializer import get_fee_charges
-from package.calculation_input.charges_reader import _create_charges_df
+from package.calculation.wholesale.charges_reader import _create_charges_df
 import pytest
